+++ src/deepskylog/AstronomyLibrary/Imports/ConstellationBoundariesImport.php
@@ -3,6 +3,7 @@
 namespace deepskylog\AstronomyLibrary\Imports;
 
 use deepskylog\AstronomyLibrary\Models\ConstellationBoundaries;
+use Illuminate\Database\Eloquent\Model;
 use Maatwebsite\Excel\Concerns\ToModel;
 use Maatwebsite\Excel\Concerns\WithCustomCsvSettings;
 
@@ -10,9 +11,9 @@ class ConstellationBoundariesImport implements ToModel, WithCustomCsvSettings
 {
     /**
      * @param  array  $row  The row of the csv file
-     * @return \Illuminate\Database\Eloquent\Model|null
+     * @return Model|array<int, Model>|null
      */
-    public function model(array $row)
+    public function model(array $row): Model|array|null
     {
         return new ConstellationBoundaries(
             [