+++ src/deepskylog/AstronomyLibrary/Imports/DeltaTImport.php
@@ -3,15 +3,16 @@
 namespace deepskylog\AstronomyLibrary\Imports;
 
 use deepskylog\AstronomyLibrary\Models\DeltaT;
+use Illuminate\Database\Eloquent\Model;
 use Maatwebsite\Excel\Concerns\ToModel;
 
 class DeltaTImport implements ToModel
 {
     /**
      * @param  array  $row  The row of the csv file
-     * @return \Illuminate\Database\Eloquent\Model|null
+     * @return Model|array<int, Model>|null
      */
-    public function model(array $row)
+    public function model(array $row): Model|array|null
     {
         return new DeltaT(
             [