+++ src/deepskylog/AstronomyLibrary/Targets/Moon.php
@@ -360,21 +360,15 @@ class Moon extends Target
 
         $equa_coords = $ecl->convertToEquatorial($nutation[3]);
 
-        // Calculate corrections for parallax
-        $pi = rad2deg(asin(6378.14 / $helio_coords[2]));
-
-        $siderial_time = Time::apparentSiderialTime($date, new GeographicalCoordinates(0.0, 0.0));
-
-        $hour_angle = (new \deepskylog\AstronomyLibrary\Coordinates\Coordinate($equa_coords->getHourAngle($siderial_time) + $geo_coords->getLongitude()->getCoordinate() * 15.0, 0, 360))->getCoordinate();
-
-        $earthsGlobe = $geo_coords->earthsGlobe($height);
-
-        $deltara = rad2deg(atan(-$earthsGlobe[1] * sin(deg2rad($pi / 3600.0)) * sin(deg2rad($hour_angle)) / (cos(deg2rad($equa_coords->getDeclination()->getCoordinate())) - $earthsGlobe[1] * sin(deg2rad($pi / 3600.0)) * sin(deg2rad($hour_angle)))));
-        $dec = rad2deg(atan((sin(deg2rad($equa_coords->getDeclination()->getCoordinate())) - $earthsGlobe[0] * sin(deg2rad($pi / 3600.0))) * cos(deg2rad($deltara / 3600.0))
-                        / (cos(deg2rad($equa_coords->getDeclination()->getCoordinate())) - $earthsGlobe[1] * sin(deg2rad($pi / 3600.0)) * cos(deg2rad($height)))));
-
-        $equa_coords->setRA($equa_coords->getRA()->getCoordinate() + $deltara);
-        $equa_coords->setDeclination($dec);
+        // Calculate corrections for parallax.
+        // The equatorial horizontal parallax in degrees.
+        $equa_coords = $this->_correctForParallax(
+            $equa_coords,
+            rad2deg(asin(6378.14 / $helio_coords[2])),
+            $date,
+            $geo_coords,
+            $height
+        );
 
         return $equa_coords;
     }